requirements.txt
Flask==3.0.0
supabase==2.3.4
google-generativeai==0.3.2
python-dotenv==1.0.0
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.1
Pillow==10.1.0
requests==2.31.0
gotrue==2.2.0
